auth-service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
python-consul
requests
mysqlclient
pika